rock5: CI release process
Build the kernel and hack the Talos `Makefile` slightly to pass
the board and optionally a specific kernel commit via a context
override. (This isn't done using the standard Talos approach for
custom kernels where you re-pack the initfs since we have a forked
installer anyway, this is faster because we can do it in a single
pack and `xz` takes forever.)

All commits get Docker images for the kernel + installer. Tags
get `metal-rock_5b-arm64.img.xz` generated from the commit image (and
the installer re-tagged with the Git tag) and attached to a GitHub
release.