rock5b: CI release process
Build the kernel and hack the Talos `Makefile` slightly to pass
the exact tag through for the context to generate the installer.
(This isn't done using the typical Talos layered approach because
we have a forked installer anyway, so this is faster.)

All commits get Docker images for the kernel + installer. Tags
get `metal-rock_5b-arm64.img.xz` generated from the commit image (and
the installer re-tagged with the Git tag) and attached to a GitHub
release.