rock5b: ignore lost+found when preserving partition
As this fork uses ext4 for the `BOOT` partition for compatibility
with the vendor U-Boot, `/lost+found` exists on that (unlike the
xfs partitions, which do not).

As a result, the archiver needs to be ignored when preserving the
filesystem or the installer will fail at upgrade to to the directory
already existing on the new partition.

For safety/sanity, any files in `/lost+found` are also ignored.
Given that `/boot` is read-only, there isn't really anything we
should ever need to manually restore from this partition.

This is intended to fix upgrades without `--preserve`.